Alt (#2)
create []Interceptor to separate handling from main

+++ syscalls/syscalls.go
@@ -7,8 +7,8 @@ import (
 
 //go:generate go run ../generate/main.go
 
-func GetName(syscallID uint64) string {
-	if name, ok := syscallNames[int(syscallID)]; ok {
+func GetName(syscallID int) string {
+	if name, ok := syscallNames[syscallID]; ok {
 		return name
 	} else {
 		return fmt.Sprintf("_%d_", syscallID)
