+++ syscalls/syscalls_test.go
@@ -7,14 +7,14 @@ import (
 
 func TestGetName(t *testing.T) {
 	syscallID := syscall.SYS_EXECVE
-	if name, expected := GetName(uint64(syscallID)), "EXECVE"; name != expected {
+	if name, expected := GetName(syscallID), "EXECVE"; name != expected {
 		t.Errorf("expected %s for %d, but got %s", expected, syscallID, name)
 	}
 }
 
 func TestGetNameFromImport(t *testing.T) {
 	syscallID := syscall.SYS_ACCT
-	if name, expected := GetName(uint64(syscallID)), "ACCT"; name != expected {
+	if name, expected := GetName(syscallID), "ACCT"; name != expected {
 		t.Errorf("expected %s for %d, but got %s", expected, syscallID, name)
 	}
 }
